fix(mdx): correctly recognize JSX
